+++ src/test/java/nextstep/subway/line/domain/LineTest.java
@@ -1,6 +1,5 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.section.domain.Section;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.Test;
